backend/requirements.txt
anyio==4.4.0
asgiref==3.8.1
async-property==0.2.2
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
daphne==4.1.2
deprecation==2.1.0
Django
django-cors-headers
django-environ
django-filter
djangorestframework
djangorestframework-simplejwt
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
itypes==1.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
Markdown==3.7
MarkupSafe==2.1.5
marshmallow==3.21.3
packaging==24.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
python-keycloak==4.3.0
requests==2.32.3
requests-toolbelt==1.0.0
service-identity==24.1.0
sniffio==1.3.1
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uuid==1.30
zope.interface==7.0.3
gunicorn
django-jazzmin